chore(deps): 升级项目依赖到最新版本
主要变更:

- reqwest 0.13: rustls-tls -> rustls, 临时移除 brotli（8.0.3 依赖冲突）

- ureq 3: 改用 Agent::config_builder 与新 Body API

- toml 1.1: build.rs 解析从 Value 改为 Table

- sha2 0.11: 新增 hex_digest 辅助函数, 替换 Write/copy 用法

- rusqlite 0.40: u64 字段改为 i64

- zip 8: ZipFile 改为泛型签名

- type-state-builder 0.5: default 属性改为真实表达式

验证: cargo test / cargo clippy / cargo fmt / check-complexity 均通过

+++ src/sync/record.rs
@@ -12,7 +12,7 @@ struct RecordData {
     package_name: String,
     version: String,
     sha256: String,
-    size: Option<u64>,
+    size: Option<i64>,
     yanked: Option<String>,
 }
 
@@ -29,8 +29,10 @@ pub async fn record_download_results(
             let yanked = fi.yanked.clone();
             async move {
                 let sha256 = hash_file_async(&dest, fi.sha256.clone()).await;
-                let size =
-                    tokio::fs::metadata(&dest).await.ok().map(|m| m.len());
+                let size = tokio::fs::metadata(&dest)
+                    .await
+                    .ok()
+                    .map(|m| m.len() as i64);
                 RecordData {
                     filename: fi.filename.clone(),
                     package_name: fi.package_name.clone(),
@@ -99,7 +101,10 @@ async fn try_hash_file(
         return None;
     }
     let sha256 = hash_file_async(&dest, fi.sha256.clone()).await;
-    let size = tokio::fs::metadata(&dest).await.ok().map(|m| m.len());
+    let size = tokio::fs::metadata(&dest)
+        .await
+        .ok()
+        .map(|m| m.len() as i64);
     Some(RecordData {
         filename: fi.filename,
         package_name: fi.package_name,
